requirements.txt
aiogram==3.7.0
apscheduler==3.10.4
python-dotenv==1.0.0
psutil==5.9.8
